Extend services proposals

+++ app/controllers/services_controller.rb
@@ -27,8 +27,11 @@ class ServicesController < ApplicationController
   def show
     @service = Service.find(params[:id])
     @my_proposal = nil
+    @accepted_proposal = nil
     if @service.user != current_user
       @my_proposal = Proposal.where(user: current_user, service: @service).first
+    else
+      @accepted_proposal = Proposal.where(service: @service, accepted: true).first
     end
   end
 
@@ -38,6 +41,11 @@ class ServicesController < ApplicationController
 
   def update
     service = Service.find(params[:id])
+    # Cancel destruction if accepted proposal exists
+    if Proposal.where(service: service, accepted: true).count > 0
+      flash[:alert] = "Can't modify a service after accepting a proposal"
+      redirect_to service_path(service.id)
+    end
     service.title = params[:service][:title]
     service.description = params[:service][:description]
     service.location = params[:service][:location]
@@ -52,6 +60,11 @@ class ServicesController < ApplicationController
 
   def destroy
     service = Service.find(params[:id])
+    # Cancel destruction if accepted proposal exists
+    if Proposal.where(service: service, accepted: true).count > 0
+      flash[:alert] = "Can't remove a service after accepting a proposal"
+      redirect_to service_path(service.id)
+    end
     service.destroy
     flash[:alert] = "Job #{service.title} has been removed"
     redirect_to services_path
@@ -59,5 +72,17 @@ class ServicesController < ApplicationController
 
   def my_services
     @services = Service.where(user: current_user)
+  end
+
+  def complete_service
+    service = Service.find(params[:id])
+    service.completed_on = DateTime.now
+    if service.save
+      flash[:alert] = "Job #{service.title} has been completed"
+      redirect_to service_path(service.id)
+    else
+      flash[:alert] = service.errors.full_messages[0]
+      redirect_to service_path(service.id)
+    end
   end
 end
